+++ src/redirect.rs
@@ -128,7 +128,7 @@ impl Policy {
         match self.inner {
             PolicyKind::Custom(ref custom) => custom(attempt),
             PolicyKind::Limit(max) => {
-                if attempt.previous.len() == max {
+                if attempt.previous.len() >= max {
                     attempt.error(TooManyRedirects)
                 } else {
                     attempt.follow()
@@ -270,6 +270,18 @@ fn test_redirect_policy_limit() {
     }
 
     previous.push(Url::parse("http://a.b.d/e/33").unwrap());
+
+    match policy.check(StatusCode::FOUND, &next, &previous) {
+        ActionKind::Error(err) if err.is::<TooManyRedirects>() => (),
+        other => panic!("unexpected {:?}", other),
+    }
+}
+
+#[test]
+fn test_redirect_policy_limit_to_0() {
+    let policy = Policy::limited(0);
+    let next = Url::parse("http://x.y/z").unwrap();
+    let previous = vec![Url::parse("http://a.b/c").unwrap()];
 
     match policy.check(StatusCode::FOUND, &next, &previous) {
         ActionKind::Error(err) if err.is::<TooManyRedirects>() => (),